[Refactor] 좌석 입퇴실 패키지 구조 변경
controller
service
repository
domain
dto
